# fix: improve the loading bot performance