--- a/Composer/packages/client/src/recoilModel/dispatchers/project.ts
+++ b/Composer/packages/client/src/recoilModel/dispatchers/project.ts
@@ -478,7 +478,7 @@ export const projectDispatcher = () => {
             // Bot creation successful
             clearInterval(timer);
             callbackHelpers.set(botOpeningMessage, response.data.latestMessage);
-            const { botFiles, projectData } = loadProjectData(response.data.result);
+            const { botFiles, projectData } = await loadProjectData(response.data.result);
             const projectId = response.data.result.id;
             if (settingStorage.get(projectId)) {
               settingStorage.remove(projectId);