--- a/Composer/packages/client/src/recoilModel/parsers/workers/luParser.worker.ts
+++ b/Composer/packages/client/src/recoilModel/parsers/workers/luParser.worker.ts
@@ -11,6 +11,7 @@ import {
   LuUpdateIntentPayload,
   LuAddIntentsPayload,
   LuAddIntentPayload,
+  LuParseAllPayload,
 } from '../types';
 const ctx: Worker = self as any;
 
@@ -50,7 +51,20 @@ interface RemoveIntentsMessage {
   payload: LuRemoveIntentsPayload;
 }
 
-type LuMessageEvent = ParseMessage | AddMessage | AddsMessage | UpdateMessage | RemoveMessage | RemoveIntentsMessage;
+type ParseAllMessage = {
+  id: string;
+  type: LuActionType.ParseAll;
+  payload: LuParseAllPayload;
+};
+
+type LuMessageEvent =
+  | ParseMessage
+  | AddMessage
+  | AddsMessage
+  | UpdateMessage
+  | RemoveMessage
+  | RemoveIntentsMessage
+  | ParseAllMessage;
 
 export const handleMessage = (msg: LuMessageEvent) => {
   let result: any = null;
@@ -58,6 +72,12 @@ export const handleMessage = (msg: LuMessageEvent) => {
     case LuActionType.Parse: {
       const { id, content, luFeatures } = msg.payload;
       result = luUtil.parse(id, content, luFeatures);
+      break;
+    }
+
+    case LuActionType.ParseAll: {
+      const { luResources, luFeatures } = msg.payload;
+      result = luResources.map(({ id, content }) => luUtil.parse(id, content, luFeatures));
       break;
     }