--- a/Composer/packages/client/src/recoilModel/parsers/workers/qnaParser.worker.ts
+++ b/Composer/packages/client/src/recoilModel/parsers/workers/qnaParser.worker.ts
@@ -2,34 +2,48 @@
 // Licensed under the MIT License.
 import * as qnaUtil from '@bfc/indexers/lib/utils/qnaUtil';
 
-import { QnAActionType } from './../types';
+import { QnAActionType, QnAParseAllPayload, QnAParsePayload } from './../types';
 const ctx: Worker = self as any;
 
-ctx.onmessage = function (msg) {
-  const { id: msgId, type, payload } = msg.data;
-  const { content, id, file, indexId } = payload;
+type ParseMessage = {
+  id: string;
+  type: QnAActionType.Parse;
+  payload: QnAParsePayload;
+};
+
+type ParseAllMessage = {
+  id: string;
+  type: QnAActionType.ParseAll;
+  payload: QnAParseAllPayload;
+};
+
+type QnaMessageEvent = ParseMessage | ParseAllMessage;
+
+export const handleMessage = (msg: QnaMessageEvent) => {
   let result: any = null;
-  try {
-    switch (type) {
-      case QnAActionType.Parse: {
-        result = qnaUtil.parse(id, content);
-        break;
-      }
-      case QnAActionType.AddSection: {
-        result = qnaUtil.addSection(file.content, content);
-        break;
-      }
-      case QnAActionType.UpdateSection: {
-        result = qnaUtil.updateSection(indexId, file.content, content);
-        break;
-      }
-      case QnAActionType.RemoveSection: {
-        result = qnaUtil.removeSection(indexId, file.content);
-        break;
-      }
+  switch (msg.type) {
+    case QnAActionType.Parse: {
+      const { id, content } = msg.payload;
+      result = qnaUtil.parse(id, content);
+      break;
     }
-    ctx.postMessage({ id: msgId, payload: result });
+
+    case QnAActionType.ParseAll: {
+      const { qnaResources } = msg.payload;
+      result = qnaResources.map(({ id, content }) => qnaUtil.parse(id, content));
+      console.log(result);
+      break;
+    }
+  }
+  return result;
+};
+
+ctx.onmessage = function (event) {
+  const msg = event.data as QnaMessageEvent;
+  try {
+    const payload = handleMessage(msg);
+    ctx.postMessage({ id: msg.id, payload });
   } catch (error) {
-    ctx.postMessage({ id: msgId, error: error.message });
+    ctx.postMessage({ id: msg.id, error: error.message });
   }
 };