## Description
**issue**
1. The lg files are parsed twice
2. composer will wait for all resources process to finish

for calendar skill, parsing time is 19s, some other tasks' (render, state set) time is 3s, the total time to finish is 22s
![image](https://user-images.githubusercontent.com/39758135/108589780-0cfd8b00-739b-11eb-8981-06e92230bb72.png)

**fix**
1. reduce one lg files' index. move all parsing operations to the worker.
2. put the lg files index to background. If we open a new dialog, check the lg file state and do a single parse if the file is not yet parsed during the loading stage.

The total time now to show the design page is about 5s
![image](https://user-images.githubusercontent.com/39758135/108590091-abd6b700-739c-11eb-9075-de8793ddc000.png)

**next step**
If this solution is ok, we can let lu and qna use the same solution during loading.
<!---
Please include a summary of the change and which issue is fixed. Please also include relevant motivation and context.

If this is a bug fix, please describe the root cause and analysis of this problem.
---->

## Task Item
refs #5708
<!---
Please include a link to the related issue. [Ex. `Closes #<issue #>`](https://help.github.com/en/articles/closing-issues-using-keywords)
---->

## Screenshots
![loadtime](https://user-images.githubusercontent.com/39758135/108590246-5cdd5180-739d-11eb-9364-4378cad02f51.gif)

<!---
Please include screenshots or gifs if your PR include UX changes.
--->